test: Extract the tests for function A into a separate file to avoid conflict
